update to 2.2
added publish / debug options for cleaner
folder del is still broken

+++ Cleaner.cs
@@ -9,7 +9,7 @@ namespace BFBB_Iso_Cleaner
         {
             if (args.Length == 0)
             {
-                string appVersion = "2.1";
+                string appVersion = "2.2";
                 Log("BFBB Iso Cleaner\nVersion: " + appVersion + "\n");
 
                 Log("Select the game path (Type it due to limited knowledge, type path with main files like sb.ini):");
@@ -18,7 +18,7 @@ namespace BFBB_Iso_Cleaner
                 var line = Console.ReadLine();
                 if (Directory.Exists(line))
                 {
-                    gamePath = Console.ReadLine();
+                    gamePath = line;
                 }
                 else
                 {
@@ -47,12 +47,32 @@ namespace BFBB_Iso_Cleaner
                     putInFolder = false;
                 }
 
-                string[] extensions = { ".sdf", ".log", ".mpl", ".lip", ".mpl~", ".scc", ".bat~", ".bat", ".out", ".lop", ".dsf", ".dlf"};
-                string deletedFilesPath = Convert.ToString(Directory.GetParent(gamePath));
+                bool publish = false;
+                Log("Is this folder for a mod publish or a debug test? (Y = publish N = debug) (Y/N): ");
+
+                var l = Console.ReadLine().ToLower();
+
+                if (l == "y")
+                {
+                    publish = true;
+                }
+                else if (l == "n")
+                {
+                    publish = false;
+                }
+                else
+                {
+                    Log("Invalid Argument! Changing to N");
+                    publish = false;
+                }
+
+                string[] debugExtensions = {".sdf", ".log", ".mpl", ".lip", ".mpl~", ".scc", ".bat~", ".bat", ".out", ".lop", ".dsf", ".dlf", ".bkp"};
+                string[] publishExtensions = {".elf", ".tpl", ".bnr"};
+                string deletedFilesPath = Convert.ToString(Directory.GetParent(gamePath)) + @"\DeletedFiles";
                 string[] fileNamesWithPaths = Directory.GetFiles(gamePath,"*",SearchOption.AllDirectories);
                 foreach (string file in fileNamesWithPaths)
                 {
-                    foreach (string ext in extensions)
+                    foreach (string ext in debugExtensions)
                     {
                         var extensionOfFile = Path.GetExtension(file);
                         if (extensionOfFile.ToLower() == ext)
@@ -67,18 +87,61 @@ namespace BFBB_Iso_Cleaner
                                 }
                                 else
                                 {
-                                    MoveFile(file, deletedFilesPath);
-                                    Log("Moved file: " + file + " To: " + deletedFilesPath);
+                                    if (File.Exists(file))
+                                    {
+                                        MoveFile(file, deletedFilesPath);
+                                        Log("Moved file: " + file + " To: " + deletedFilesPath);
+                                    }
                                 }
                             }
                             else
                             {
-                                DeleteFile(file);
-                                Log("Deleted file: " + file);
+                                if (File.Exists(file))
+                                {
+                                    DeleteFile(file);
+                                    Log("Deleted file: " + file);
+                                }
                             }
                         }
                     }
                 }
+
+                if (publish == true)
+                {
+                    foreach (string file in fileNamesWithPaths)
+                    {
+                        foreach (string ext in publishExtensions)
+                        {
+                            var extensionOfFile = Path.GetExtension(file);
+                            if (extensionOfFile.ToLower() == ext)
+                            {
+                                if (putInFolder == true)
+                                {
+                                    if (Directory.Exists(deletedFilesPath) == false)
+                                    {
+                                        Log("Could not find deleted path folder!");
+                                        Directory.CreateDirectory(deletedFilesPath);
+                                        Log("Created deleted path folder!");
+                                    }
+                                    else
+                                    {
+                                        MoveFile(file, deletedFilesPath);
+                                        Log("Moved file: " + file + " To: " + deletedFilesPath);
+                                    }
+                                }
+                                else
+                                {
+                                    DeleteFile(file);
+                                    Log("Deleted file: " + file);
+                                }
+                            }
+                        }
+                    }
+                }
+
+                Log("Iso Sucessfully Cleaned! Press any key to close...");
+                Console.ReadKey();
+                Environment.Exit(0);
             }
         }
 
@@ -86,7 +149,10 @@ namespace BFBB_Iso_Cleaner
         {
             try
             {
-                File.Delete(path);
+                if (File.Exists(path) == true)
+                {
+                    File.Delete(path);
+                }
             }
             catch (Exception e)
             {
@@ -100,7 +166,10 @@ namespace BFBB_Iso_Cleaner
         {
             try
             {
-                File.Move(path, dest);
+                if (File.Exists(path))
+                {
+                    File.Move(path, dest);
+                }
             }
             catch (Exception e)
             {
